Add files via upload
This code is written in Python. It utilizes the PyQt5 library for building GUI applications, matplotlib for plotting graphs, seaborn for data visualization, pandas for data manipulation, numpy for numerical computations, and scikit-learn for linear regression modeling. Additionally, it uses scipy for statistical computations and seaborn for data visualization. The code appears to be creating a GUI application for analyzing and visualizing global economy indicators data.